Tournaments construction and socket code restructure
